requirements.txt
torch
transformers
evaluate
seqeval
optuna
Pillow